pyproject.toml
[tool.poetry]
name = "booktrack-fastapi"
version = "0.1.0"
description = ""
authors = ["Rafaella Duarte <rafaella.d.d.carvalho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.121.0"}
sqlalchemy = "^2.0.44"
pydantic-settings = "^2.12.0"
alembic = "^1.17.1"
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
pandas = "^2.3.3"
aiosqlite = "^0.22.0"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.4,<0.15.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.ruff]
line-length = 100
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev booktrack_fastapi/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=booktrack_fastapi -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"